+++ tiledb/sm/query/query.cc
@@ -704,6 +704,7 @@ Status Query::check_subarray(const void* subarray) const {
     case Datatype::STRING_UTF32:
     case Datatype::STRING_UCS2:
     case Datatype::STRING_UCS4:
+    case Datatype::ANY:
       // Not supported domain type
       assert(0);
       break;