+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Added string type support (ASCII, UTF-8, UTF-16, UTF-32, UCS-2, UCS-4)
+* Added `TILEDB_ANY` datatype
 
 ## API additions